refactor: split package by domain
