+++ chaekit-spring/src/main/java/qwerty/chaekit/domain/group/activity/discussion/Discussion.java
@@ -10,7 +10,7 @@ import qwerty.chaekit.domain.BaseEntity;
 import qwerty.chaekit.domain.group.activity.discussion.highlight.DiscussionHighlight;
 import qwerty.chaekit.domain.group.activity.Activity;
 import qwerty.chaekit.domain.group.activity.discussion.comment.DiscussionComment;
-import qwerty.chaekit.domain.highlight.entity.Highlight;
+import qwerty.chaekit.domain.highlight.Highlight;
 import qwerty.chaekit.domain.member.user.UserProfile;
 
 import java.util.ArrayList;
