+++ chaekit-spring/src/main/java/qwerty/chaekit/domain/highlight/comment/repository/HighlightCommentJpaRepository.java
@@ -1,7 +1,7 @@
-package qwerty.chaekit.domain.highlight.repository.comment;
+package qwerty.chaekit.domain.highlight.comment.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import qwerty.chaekit.domain.highlight.entity.comment.HighlightComment;
+import qwerty.chaekit.domain.highlight.comment.HighlightComment;
 
 public interface HighlightCommentJpaRepository extends JpaRepository<HighlightComment, Long> {
 }
