+++ chaekit-spring/src/main/java/qwerty/chaekit/domain/highlight/reaction/ReactionType.java
@@ -1,4 +1,4 @@
-package qwerty.chaekit.domain.highlight.entity.reaction;
+package qwerty.chaekit.domain.highlight.reaction;
 
 public enum ReactionType{
     GREAT("👍"),
