+++ chaekit-spring/src/main/java/qwerty/chaekit/domain/highlight/reaction/repository/HighlightReactionRepositoryImpl.java
@@ -1,9 +1,9 @@
-package qwerty.chaekit.domain.highlight.repository.reaction;
+package qwerty.chaekit.domain.highlight.reaction.repository;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-import qwerty.chaekit.domain.highlight.entity.reaction.HighlightReaction;
-import qwerty.chaekit.domain.highlight.entity.reaction.ReactionType;
+import qwerty.chaekit.domain.highlight.reaction.HighlightReaction;
+import qwerty.chaekit.domain.highlight.reaction.ReactionType;
 
 import java.util.List;
 import java.util.Optional;
