+++ chaekit-spring/src/main/java/qwerty/chaekit/domain/highlight/repository/HighlightRepositoryImpl.java
@@ -9,8 +9,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 import qwerty.chaekit.domain.group.ReadingGroup;
 import qwerty.chaekit.domain.group.activity.Activity;
-import qwerty.chaekit.domain.highlight.entity.Highlight;
-import qwerty.chaekit.domain.highlight.entity.QHighlight;
+import qwerty.chaekit.domain.highlight.Highlight;
+import qwerty.chaekit.domain.highlight.QHighlight;
 import qwerty.chaekit.domain.member.user.UserProfile;
 
 import java.util.List;
