+++ chaekit-spring/src/main/java/qwerty/chaekit/dto/highlight/HighlightSummaryResponse.java
@@ -1,11 +1,7 @@
 package qwerty.chaekit.dto.highlight;
 
 import lombok.Builder;
-import qwerty.chaekit.domain.group.activity.discussion.Discussion;
-import qwerty.chaekit.domain.highlight.entity.Highlight;
-import qwerty.chaekit.dto.group.activity.discussion.DiscussionSummaryResponse;
-
-import java.util.List;
+import qwerty.chaekit.domain.highlight.Highlight;
 
 @Builder
 public record HighlightSummaryResponse(
