+++ chaekit-spring/src/main/java/qwerty/chaekit/service/highlight/comment/HighlightCommentService.java
@@ -4,12 +4,12 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import qwerty.chaekit.domain.highlight.entity.Highlight;
-import qwerty.chaekit.domain.highlight.entity.comment.HighlightComment;
-import qwerty.chaekit.domain.highlight.entity.reaction.HighlightReaction;
+import qwerty.chaekit.domain.highlight.Highlight;
+import qwerty.chaekit.domain.highlight.comment.HighlightComment;
+import qwerty.chaekit.domain.highlight.reaction.HighlightReaction;
 import qwerty.chaekit.domain.highlight.repository.HighlightRepository;
-import qwerty.chaekit.domain.highlight.repository.comment.HighlightCommentRepository;
-import qwerty.chaekit.domain.highlight.repository.reaction.HighlightReactionRepository;
+import qwerty.chaekit.domain.highlight.comment.repository.HighlightCommentRepository;
+import qwerty.chaekit.domain.highlight.reaction.repository.HighlightReactionRepository;
 import qwerty.chaekit.domain.member.user.UserProfile;
 import qwerty.chaekit.domain.member.user.UserProfileRepository;
 import qwerty.chaekit.dto.highlight.comment.HighlightCommentRequest;
